mega refactor
