+++ .gitignore
@@ -14,7 +14,7 @@ gradle.properties
 /*.json*
 *.tmp
 /.vscode
-launcher/
+/launcher
 /ViaLoader/
 dist/
 /*.spec
