+++ src/launcher/main.py
@@ -1,0 +1,33 @@
+import time
+
+import github_api
+import launcher
+from launch_config import LaunchConfig, read_launch_config_file
+from setup import setup_execute
+from update_launcher import update_launcher_exec
+from update_zenith import update_zenith_exec
+
+config = LaunchConfig()
+api = github_api.GitHubAPI(config)
+launcher.config = config
+launcher.api = api
+
+
+def main():
+    while True:
+        json_data = read_launch_config_file()
+        if json_data is None:
+            print("launch_config.json not found, running setup.")
+            setup_execute()
+            json_data = read_launch_config_file()
+        config.load_launch_config_data(json_data)
+        config.validate_launch_config()
+        update_launcher_exec(config, api)
+        update_zenith_exec(config, api)
+        launcher.launcher_exec(config, api)
+        print("Restarting in 3 seconds...")
+        time.sleep(3)
+
+
+if __name__ == "__main__":
+    main()
