+++ src/launcher/utils.py
@@ -1,0 +1,6 @@
+import sys
+
+
+def critical_error(message):
+    print("CRITICAL: " + message)
+    sys.exit(69)
